refactor: reorganize tests following Rust best practices
- Move unit tests from tests/ directory to their respective source files
- Parser tests → src/pubmed/parser.rs #[cfg(test)] module
- Model utility tests → src/pubmed/models.rs #[cfg(test)] module
- Query builder tests → src/pubmed/query.rs #[cfg(test)] module (already existed)
- Keep integration tests in tests/test_mesh_integration.rs
- Update all doc test imports from query:: to pubmed:: after module refactoring

Benefits:
- Unit tests are now co-located with the code they test
- Integration tests are separate and properly marked with #[ignore]
- Better organization following Rust conventions
- All tests pass (62 unit + 65 doc tests)
